Tighten production safety checks

+++ core/graph_engine.py
@@ -23,7 +23,7 @@ class AttackGraph:
         for src, dst, data in self.graph.edges(data=True):
             print(f"{src} --[{data.get('label', 'link')}]-> {dst}")
 
-    def export_to_neo4j(self, uri=None, user=None, password=None):
+    def export_to_neo4j(self, uri=None, user=None, password=None, clear=False):
         """
         Exporte le graphe vers Neo4j.
         
@@ -41,15 +41,16 @@ class AttackGraph:
         
         print("[*] Export vers Neo4j...")
         neo = Graph(uri, auth=(user, password))
-        neo.delete_all()
+        if clear:
+            neo.delete_all()
 
         for node, attrs in self.graph.nodes(data=True):
-            neo.merge(Node(attrs.get("type", "Node"), name=node))
+            neo.merge(Node(attrs.get("type", "Node"), name=node), attrs.get("type", "Node"), "name")
 
         for src, dst, data in self.graph.edges(data=True):
             a = Node(self.graph.nodes[src].get("type", "Node"), name=src)
             b = Node(self.graph.nodes[dst].get("type", "Node"), name=dst)
             rel = Relationship(a, data.get("label", "LINKS_TO"), b)
-            neo.merge(rel)
+            neo.merge(rel, self.graph.nodes[src].get("type", "Node"), "name")
 
         print("[+] Export terminé.")
